Refactor chain finder to its own file
