+++ common/design_utils.cc
@@ -74,7 +74,8 @@ void print_utilisation(const Context *ctx)
 }
 
 // Connect a net to a port
-void connect_port(const Context *ctx, NetInfo *net, CellInfo *cell, IdString port_name) {
+void connect_port(const Context *ctx, NetInfo *net, CellInfo *cell, IdString port_name)
+{
     PortInfo &port = cell->ports.at(port_name);
     NPNR_ASSERT(port.net == nullptr);
     port.net = net;
